refactor: move all the redundant Fragment codes into the BaseFragment
