+++ app/src/main/java/com/simpleapp/challenge/ui/userlist/UserListFragment.kt
@@ -7,7 +7,6 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.core.content.ContextCompat
-import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
@@ -17,13 +16,14 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.simpleapp.challenge.R
 import com.simpleapp.challenge.databinding.FragmentUserListBinding
+import com.simpleapp.challenge.ui.base.BaseFragment
 import com.simpleapp.challenge.ui.login.LoginActivity
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
-class UserListFragment : Fragment() {
+class UserListFragment : BaseFragment() {
 
   lateinit var binding: FragmentUserListBinding
   val viewModel: UserListViewModel by activityViewModels()
@@ -56,6 +56,10 @@ class UserListFragment : Fragment() {
 
   }
 
+  override fun getToolbarTitle(): String {
+    return getString(R.string.title_activity_user_list)
+  }
+
   override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
     binding = FragmentUserListBinding.inflate(inflater, container, false)
     binding.viewModel = viewModel
@@ -66,7 +70,7 @@ class UserListFragment : Fragment() {
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
 
-    setupToolbar()
+    setHomeButtonEnabled(false)
     setupRecyclerView()
   }
 
@@ -83,15 +87,6 @@ class UserListFragment : Fragment() {
     viewModel.userList.observe(viewLifecycleOwner, { list ->
       adapter.updateItemsList(list)
     })
-  }
-
-  private fun setupToolbar() {
-    (requireActivity() as UserListActivity).supportActionBar?.apply {
-      setDisplayHomeAsUpEnabled(false)
-      setHomeButtonEnabled(false)
-      setHasOptionsMenu(true)
-      setTitle(R.string.title_activity_user_list)
-    }
   }
 
   private fun navigateToLogin() {
